requirements.txt
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-aiplatform==1.74.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.19.0
googleapis-common-protos==1.66.0
Jinja2==3.1.4
proto-plus==1.25.0
protobuf==5.29.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
urllib3==2.2.3
uvicorn==0.34.0
fastapi==0.115.8